map_checker.c created, main file cleaned
